requirements.txt
# Core API clients
anthropic==0.42.0
openai==1.59.4

# Data handling and datasets
datasets==3.2.0
names-dataset==3.1.0

# Configuration and utilities
PyYAML==6.0.2
tqdm==4.67.1

# Scientific computing (for potential mathematical operations)
numpy>=1.21.0

# HTTP requests (for API calls if needed)
requests>=2.25.0

# JSON handling enhancements
ujson>=5.0.0

# Logging and debugging
colorlog>=6.0.0

# Optional: For better error handling and retries
tenacity>=8.0.0

# Optional: For data analysis and statistics
pandas>=1.3.0

# Optional: For better terminal output formatting
rich>=13.0.0
